Revert SSR bullshit
